CLEAN (moved odometry to own class)
